ApiException + UserNotFoundApiException

+++ src/Controller/UserApiController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Entity\UserEditDTO;
 use App\Entity\UserSaveDTO;
+use App\Exception\UserNotFoundApiException;
 use App\Repository\UserRepository;
 use App\Utils\ApiResponse;
 use Exception;
@@ -28,12 +29,16 @@ class UserApiController extends AbstractController
         return $this->json(ApiResponse::get($users));
     }
 
+
     /**
      * @throws Exception
      */
     #[Route('/{id}', name: 'app_user_api_index', methods: ['GET'])]
-    public function one(User $user, UserRepository $userRepository): Response
+    public function one(int $id, UserRepository$userRepository ): Response
     {
+        $user = $userRepository->findOneBy([ "id" => $id ]);
+        if ($user === null) throw new UserNotFoundApiException();
+
         if ($user) {
             return $this->json(ApiResponse::get($user));
         }
@@ -67,8 +72,11 @@ class UserApiController extends AbstractController
      * @throws Exception
      */
     #[Route('/{id}', name: 'app_user_api_edit', methods: ['PUT', 'PATCH'])]
-    public function edit(Request $request, User $user, SerializerInterface $serializer, UserRepository $userRepository): Response
+    public function edit(Request $request, int $id, SerializerInterface $serializer, UserRepository $userRepository): Response
     {
+        $user = $userRepository->findOneBy([ "id" => $id ]);
+        if ($user === null) throw new UserNotFoundApiException();
+
         /** @var UserEditDTO $body */
         $body = $serializer->deserialize($request->getContent(), UserEditDTO::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $user]);
 
@@ -91,8 +99,11 @@ class UserApiController extends AbstractController
      * @throws Exception
      */
     #[Route('/{id}', name: 'app_user_api_delete', methods: ['DELETE'])]
-    public function delete(User $user, UserRepository $userRepository): Response
+    public function delete(int $id, UserRepository $userRepository): Response
     {
+        $user = $userRepository->findOneBy([ "id" => $id ]);
+        if ($user === null) throw new UserNotFoundApiException();
+
         $userRepository->remove($user, true);
 
         return $this->json(ApiResponse::get(null, Response::HTTP_NO_CONTENT));
